requirements.txt
Django==5.1.1
black==24.8.0
flake8==7.1.1
isort==5.13.2
pre-commit==4.0.1
djangorestframework==3.15.2
django-filter==24.3
django-storages==1.14.4
boto3==1.35.29
django-environ==0.11.2